+++ flake.nix
@@ -93,6 +93,15 @@
     };
   in mkConfigurations [
     {
+      host = "delta";
+      user = "enzime";
+      system = "x86_64-linux";
+      nixos = true;
+      modules = builtins.attrValues {
+        inherit (modules) fonts gnome i3;
+      };
+    }
+    {
       host = "phi";
       user = "enzime";
       system = "x86_64-linux";